Add perft 7

+++ src/main/java/com/kelseyde/calvin/tuning/perft/PerftService.java
@@ -12,12 +12,12 @@ public class PerftService {
 
     private final MoveGenerator moveGenerator = new MoveGenerator();
 
-    public int perft(Board board, int depth) {
+    public long perft(Board board, int depth) {
         List<Move> moves = moveGenerator.generateMoves(board);
         if (depth == 1) {
             return moves.size();
         }
-        int totalMoveCount = 0;
+        long totalMoveCount = 0;
         for (Move move : moves) {
             board.makeMove(move);
             totalMoveCount += perft(board, depth - 1);
